Reference : Improve tracking of edits.
Prior to this, an edit was considered to be a plug value which was not at the default. This meant it was impossible for the end user of a reference to set a plug to the default value and expect it to stay there if the publisher of the reference changed the default value. We now consider an edit to be any value which has been explicitly set after loading the reference, even if the value has been set back to the default. We expose the edited status for plugs through a new `Reference::hasEdit()` method.