Bindings : Avoid boost::signals::detail::unusable
Since it is an implementation detail we should not have been using it anyway, and it no longer exists in boost::signals2.

The reason we were using it is because the SlotCallers would otherwise be calling `boost::python::extract<void>()` when the return type of a signal was void, and `extract<void>()` doesn't compile. We now use a simple wrapper for `extract<>` that is specialised for `void`.