Remove GafferRenderMan
The original idea behind this module was that we'd eventually support both 3Delight and PRMan using the same basic set of nodes. Since that time, the two renderers have diverged so much from the original RenderMan standard that this is no longer a viable approach. 3Delight now has an entirely new NSI API for scene description and uses OSL shaders exclusively. RenderMan has an entirely new RIS architecture with C++ shaders and partial support for OSL.

Removing this gets us one step closer to removing the legacy export code based around IECore::Renderer, and moving exclusively to using IECoreScenePreview::Renderer. It also paves the way for creating a new GafferDelight module which supports all the interactive rendering goodness of the NSI API, and for eventually creating a GafferPRMan module to do the same for PRMan 22 and beyond.