GafferOSL : Add OSLLight
Since OSL doesn't define light shaders, this is implemented as a surface shader assigned to simple geometry.

Implemented internally using OSLShader for all shader loading and network building. It has always been odd that we haven't taken this approach for the other Light subclasses, and have duplicated a bunch of code instead. Perhaps in the future we can refactor so that all lights work this way.