LocalDispatcherTest : Use `self.temporaryDirectory()`
All unit tests are meant to use this as a convenient place to create temporary files that are guaranteed to be removed on tear down. This file escaped being tidied in the past because a naive attempt at tidying it caused crashes at shutdown. This appears to have been because the `create()` function registered in `setUp()` would have `self` baked into it, and then be stored in static storage on the C++ side. This could lead to attempts to destroy Python objects after Python itself was shutdown. There's no need to register the test dispatcher anyway, since it isn't used outside this file and it just pollutes the global list, so we now just use a simple `__createLocalDispatcher()` function instead.