LocalDispatcher : Avoid deprecated `jobDirectory()` method
Also remove extraneous arguments to `Job.__init__()` - everything else is derived from the dispatcher and context inside the constructor, so I don't see any reason to split the rest of the logic out into the caller. Note that I haven't marked this as a breaking change, because the `Job` constructor is only for the use of the dispatcher itself.