Application : Enforce `-threads` argument more effectively
The old `task_scheduler_init()` call cannot stop someone creating a new arena, which will max out the thread count again. And with some versions of TBB we are forced to create a new arena in ValuePlug.cpp as a "poor man's task isolation". The `global_control` class does what it says on the tin, and limits the total thread count regardless of what other code does with arenas and task_scheduler_inits.