TaskMutex : Support recursion
Need to support this in a form where `execute()` doesn't make an arena. Also need to document it properly.

NO. I THINK THIS IS A DEAD END. FOR VALUEPLUG, WE WOULD NEED DIFFERENT THREADS TO BE ABLE TO REACQUIRE THE MUTEX, AND THAT WOULD LEAD TO CLASHING WRITES TO THE CACHEENTRY. UNLESS PERHAPS WE CAN QUERY THE FACT THAT WE HAVE A BORROWED/REENTRANT LOCK, AND USE THAT IN THE CACHE TO AVOID WRITES FOR ALL BUT THE OUTER CALL?